refactor(redis): update expire methods in Redis primitives to use i64 for seconds parameter for consistency

+++ crates/adapter/src/redis/primitives/bitmap.rs
@@ -239,7 +239,7 @@ impl RedisBitmap {
     /// Sets the TTL of a bitmap in seconds
     pub fn expire(&self, key: &str, seconds: u64) -> RedisResult<bool> {
         let mut conn = self.acquire_connection()?;
-        let result: i32 = conn.expire(key, seconds as usize)?;
+        let result: i32 = conn.expire(key, seconds as i64)?;
         Ok(result == 1)
     }
